requirements.txt
streamlit==1.28.0
openai-whisper==20231106
transformers==4.36.0
torch==2.1.0
soundfile==0.12.1